requirements.txt
altgraph
future
macholib
pefile
Pillow
PyInstaller
PyQt6
PyQt6-sip